% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{add_data_layers_to_map}
\alias{add_data_layers_to_map}
\title{Add shapefile data, and region-specific values to make Choropleth using highchartr}
\usage{
add_data_layers_to_map(
  hc,
  map_data,
  data_series_options = list(),
  color_options = list(),
  plot_options = list(),
  title_options = list(),
  subtitle_options = list(),
  caption_options = list(),
  tooltip_options = list(),
  legend_options = list()
)
}
\arguments{
\item{hc}{A highchart object.}

\item{map_data}{Data.frame. Must have entries 'id', 'name', 'path', 'value'.}

\item{data_series_options}{Optional: 'borderColor', 'borderWidth', 'dataLabels'.}

\item{color_options}{Essential: 'dataClasses', 'colors'. Optional: 'dataClassColor'.}

\item{plot_options}{Optional: 'clickEvent'.}

\item{title_options}{Essential: 'title'. Optional: 'margin', 'widthAdjust'.}

\item{subtitle_options}{Essential: 'subtitle'.}

\item{caption_options}{Essential: 'text'. Optional: 'useHTML', 'margin'.}

\item{tooltip_options}{Essential: 'pointFormat'. Optional: 'headerFormat', 'useHTML'.}

\item{legend_options}{Optional: 'reversed', 'enabled', 'layout', 'align', 'verticalAlign',
'itemMarginBottom', 'itemMarginTop', 'margin', 'titleText'.}
}
\value{
A highchart object.
}
\description{
Each of the \verb{*_options} arguments should be a list. These lists override default values for
various parts of the highchart object that is returned. There are some entries that must be
present, and a number of entries that can be absent (if absent they assume default values).
See the individual parameter entries.
}
