% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{create_base_map}
\alias{create_base_map}
\title{Create an empty highchart map object into which data layers can be added}
\usage{
create_base_map(hc_thm, export_options = list())
}
\arguments{
\item{hc_thm}{Highchart theme object, as would be used in \code{hc_add_theme()}.}

\item{export_options}{List. Options for controlling features/presentation of the image
download interface for the maps. See \code{viz_export()}.}
}
\value{
\code{highchart} htmlwidget
}
\description{
This includes functionality for exporting a map as an image
}
