% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Temporary function, until we have some real widgets in the package}
\usage{
hello(x = "World")
}
\arguments{
\item{x}{Scalar character. A name to be greeted.}
}
\description{
Temporary function, until we have some real widgets in the package
}
