% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_map2}
\alias{us_map2}
\title{us_map2
list containing shapfile data for 'us_map2'}
\format{
An object of class \code{list} of length 51.
}
\usage{
us_map2
}
\description{
us_map2
list containing shapfile data for 'us_map2'
}
\keyword{datasets}
