% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_map3}
\alias{us_map3}
\title{us_map3
list containing shapefile data for 'us_map3'}
\format{
An object of class \code{list} of length 59.
}
\usage{
us_map3
}
\description{
us_map3
list containing shapefile data for 'us_map3'
}
\keyword{datasets}
