% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{viz_export}
\alias{viz_export}
\title{Add features for downloading figures to a highchart graphic}
\usage{
viz_export(hc, export_options = list())
}
\arguments{
\item{hc}{A highchart object}

\item{export_options}{List of options for controlling how exporting works}
}
\description{
Add features for downloading figures to a highchart graphic
}
